feat: add makeTable method
