+++ src/ux/table.ts
@@ -5,6 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { TableOptions } from '@oclif/table';
+import { env } from '@salesforce/kit';
 
 type Column<T extends Record<string, unknown>> = {
   extended: boolean;
@@ -55,4 +56,49 @@ export function convertToNewTableAPI<T extends Record<string, unknown>>(
   ) as Array<Record<string, unknown>>;
 
   return { data: d, title: options?.title, borderStyle: 'headers-only-with-underline', columns: cols };
+}
+
+export function getDefaults<T extends Record<string, unknown>>(options: TableOptions<T>): Partial<TableOptions<T>> {
+  const borderStyles = [
+    'all',
+    'headers-only-with-outline',
+    'headers-only-with-underline',
+    'headers-only',
+    'horizontal-with-outline',
+    'horizontal',
+    'none',
+    'outline',
+    'vertical-with-outline',
+    'vertical',
+  ];
+
+  const defaultStyle = 'vertical-with-outline';
+  const determineBorderStyle = (): TableOptions<T>['borderStyle'] => {
+    const envVar = env.getString('SF_TABLE_BORDER_STYLE', defaultStyle);
+    if (borderStyles.includes(envVar)) {
+      return envVar as TableOptions<T>['borderStyle'];
+    }
+
+    return defaultStyle;
+  };
+
+  const overflowOptions = ['wrap', 'truncate', 'truncate-middle', 'truncate-start', 'truncate-end'];
+  const determineOverflow = (): TableOptions<T>['overflow'] => {
+    const envVar = env.getString('SF_TABLE_OVERFLOW');
+    if (envVar && overflowOptions.includes(envVar)) {
+      return envVar as TableOptions<T>['overflow'];
+    }
+
+    return options.overflow;
+  };
+
+  return {
+    borderStyle: determineBorderStyle(),
+    noStyle: env.getBoolean('SF_NO_TABLE_STYLE', false),
+    headerOptions: {
+      ...options.headerOptions,
+      formatter: 'capitalCase',
+    },
+    overflow: determineOverflow(),
+  };
 }
